Flatten controllers folder structure
- Remove duplicate Controllers/BankoApi/Controllers/ nesting
- Settings: BankoApi.Controllers.Settings
- Transactions: BankoApi.Controllers.Transactions (typo fixed)
- Users: BankoApi.Controllers.Users
- Shared: BankoApi.Controllers.Shared + .DTO
- GoCardless: untouched
- Update all usings across project and tests
- No code/behavior changes, only namespace and folder moves
