--- BankoApi/Controllers/BankoApi/Controllers/GenericResponses/ErrorResponse.cs
@@ -1,6 +1,0 @@
-namespace BankoApi.Controllers.BankoApi.Controllers.Responses;
-
-public class ErrorResponse
-{
-    public string Message { get; set; }
-}
