--- BankoApi/Controllers/BankoApi/Controllers/SettingsController/Responses/BankAuthorizationErrorMessages.cs
@@ -1,8 +1,0 @@
-﻿namespace BankoApi.Controllers.BankoApi.Controllers.SettingsController.Responses
-{
-    public enum BankAuthorizationErrorMessages
-    {
-        SomethingWentWrong,
-        NoAuthorizationFound
-    }
-}
