--- BankoApi/Controllers/BankoApi/Controllers/SettingsController/Responses/GetExpenseTagsResponse.cs
@@ -1,8 +1,0 @@
-using BankoApi.Data.Dao;
-
-namespace BankoApi.Controllers.BankoApi.Controllers.SettingsController.Responses;
-
-public class GetExpenseTagsResponse
-{
-    public List<ExpenseTag> ExpenseTags { get; set; }
-}
