--- BankoApi/Controllers/BankoApi/Controllers/SettingsController/Responses/UpsertExpenseTagResponse.cs
@@ -1,8 +1,0 @@
-using BankoApi.Data.Dao;
-
-namespace BankoApi.Controllers.BankoApi.Controllers.SettingsController.Responses;
-
-public class UpsertExpenseTagResponse
-{
-    public ExpenseTag ExpenseTag { get; set; }
-}
