--- BankoApi/Controllers/BankoApi/Controllers/TransacrionsController/Requests/UpdateNoteRequest.cs
@@ -1,6 +1,0 @@
-namespace BankoApi.Controllers.BankoApi.Controllers.TransacrionsController.Requests;
-
-public class UpdateNoteRequest
-{
-    public string Note { get; set; }
-}
