--- BankoApi/Controllers/BankoApi/DTO/UserDto.cs
@@ -1,11 +1,0 @@
-namespace BankoApi.Controllers.BankoApi.DTO;
-
-public struct UserDto
-{
-    public string Email { get; init; }
-    public string Password { get; init; }
-    public string? FullName { get; init; }
-    public string? Address { get; init; }
-    public string? PhoneNumber { get; init; }
-    public bool? ConsentGiven { get; init; }
-}
