+++ BankoApi/Controllers/Settings/Requests/UpdateExpenseTagRequest.cs
@@ -1,0 +1,7 @@
+namespace BankoApi.Controllers.Settings.Requests;
+
+public class UpdateExpenseTagRequest
+{
+    public string? ExpenseTagId { get; set; }
+    public string TransactionId { get; set; }
+}
