+++ BankoApi/Controllers/Settings/Responses/BankAuthorizationErrorMessages.cs
@@ -1,0 +1,8 @@
+namespace BankoApi.Controllers.Settings.Responses
+{
+    public enum BankAuthorizationErrorMessages
+    {
+        SomethingWentWrong,
+        NoAuthorizationFound
+    }
+}
