+++ BankoApi/Controllers/Settings/Responses/GetExpenseTagsResponse.cs
@@ -1,0 +1,8 @@
+using BankoApi.Data.Dao;
+
+namespace BankoApi.Controllers.Settings.Responses;
+
+public class GetExpenseTagsResponse
+{
+    public List<ExpenseTag> ExpenseTags { get; set; }
+}
