+++ BankoApi/Controllers/Settings/Responses/UpsertExpenseTagResponse.cs
@@ -1,0 +1,8 @@
+using BankoApi.Data.Dao;
+
+namespace BankoApi.Controllers.Settings.Responses;
+
+public class UpsertExpenseTagResponse
+{
+    public ExpenseTag ExpenseTag { get; set; }
+}
