+++ BankoApi/Controllers/Shared/DTO/UserDto.cs
@@ -1,0 +1,11 @@
+namespace BankoApi.Controllers.Shared.DTO;
+
+public struct UserDto
+{
+    public string Email { get; init; }
+    public string Password { get; init; }
+    public string? FullName { get; init; }
+    public string? Address { get; init; }
+    public string? PhoneNumber { get; init; }
+    public bool? ConsentGiven { get; init; }
+}
