+++ BankoApi/Controllers/Shared/ErrorResponse.cs
@@ -1,0 +1,6 @@
+namespace BankoApi.Controllers.Shared;
+
+public class ErrorResponse
+{
+    public string Message { get; set; }
+}
