+++ BankoApi/Controllers/Transactions/Requests/UpdateNoteRequest.cs
@@ -1,0 +1,6 @@
+namespace BankoApi.Controllers.Transactions.Requests;
+
+public class UpdateNoteRequest
+{
+    public string Note { get; set; }
+}
